[front] refactor: simplify api file structure
- removed
removed `auth.ts` and transfered to `client/src/api/auth/index.ts`

- added
added point api request services.
`client/src/api/point/index.ts`
`client/src/api/point/history.ts`
